+++ src/structures/element.rs
@@ -13,11 +13,27 @@ use crate::{
     selections::select::{parse_query, Select},
 };
 
+/*/// ## Methods for working with elements in the System.
+impl System {
+    /// Guess elements of the atoms in the system based on the provided `SupportedElements` file.
+    pub fn guess_elements(&mut self, elements: SupportedElements) -> Result<(), ElementError> {
+        // only elements that actually exist in the system 
+        // will be included in the `SupportedElements` structure associated with the system.
+        let mut new_elements = Vec::new();
+
+        for atom in self.atoms_iter_mut() {
+
+        }
+
+        
+    }
+}*/
+
 /// Contains information about all elements that can occur in the system.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 #[allow(unused)]
-pub struct ElementsProperties {
-    /// all supported elements
+pub struct SupportedElements {
+    /// All supported elements
     elements: Vec<Element>,
 }
 
@@ -25,45 +41,142 @@ pub struct ElementsProperties {
 #[derive(Debug, Clone)]
 #[allow(unused)]
 pub struct Element {
-    /// name of the element
+    /// Name of the element
     name: String,
-    /// symbol of the element
+    /// Symbol of the element
     symbol: Option<String>,
     /// `Select` structure identifying the atoms of this element
     select: Option<Select>,
-    /// atomic mass of the element in amu (daltons)
+    /// Atomic mass of the element in amu (daltons)
     mass: Option<f32>,
-    /// van der Waals radius of the atom in nm
+    /// Van der Waals radius of the atom in nm
     vdw: Option<f32>,
-    /// expected maximal number of bonds
+    /// Expected maximal number of bonds
     expected_max_bonds: Option<u8>,
 }
 
-impl Default for ElementsProperties {
-    /// Construct a default `ElementsProperties` structure.
+impl Element {
+    /// Get name of the element.
+    pub fn get_name(&self) -> &str {
+        &self.name
+    }
+
+    /// Get symbol of the element.
+    pub fn get_symbol(&self) -> Option<&str> {
+        self.symbol.as_deref()
+    }
+
+    /// Get select tree of the element.
+    pub fn get_select(&self) -> Option<&Select> {
+        self.select.as_ref()
+    }
+
+    /// Get mass of the element.
+    pub fn get_mass(&self) -> Option<f32> {
+        self.mass
+    }
+
+    /// Get van der Waals radius of the element.
+    pub fn get_vdw(&self) -> Option<f32> {
+        self.vdw
+    }
+
+    /// Get the expected maximal number of bonds.
+    pub fn get_expected_max_bonds(&self) -> Option<u8> {
+        self.expected_max_bonds
+    }
+
+}
+
+impl Default for SupportedElements {
+    /// Construct a default `SupportedElements` structure.
+    /// The structure will contain default information about elements
+    /// that are supported and recognized by the `groan_rs` library.
     ///
-    /// This function parses YAML content from `src/config/elements.yaml`
+    /// ## Notes
+    /// - This function parses YAML content from `src/config/elements.yaml`
     /// which is included in the `groan_rs` library at compile time.
-    ///
-    /// This is a relatively slow operation and there is no reason to call it multiple times in a program!
+    /// - This is a relatively slow operation and there is no reason to call it multiple times in a program!
+    /// If you need to use the `SupportedElements` structure for multiple systems, clone it.
     fn default() -> Self {
         let yaml = include_str!("../config/elements.yaml");
 
-        ElementsProperties::new_from_string(yaml)
-            .expect("FATAL GROAN ERROR | ElementsProperties::default | Default `elements.yaml` file could not be read or parsed.")
+        SupportedElements::new_from_string(yaml)
+            .expect("FATAL GROAN ERROR | SupportedElements::default | Default `elements.yaml` file could not be read or parsed.")
     }
 }
 
-impl ElementsProperties {
-    /// Construct a new `ElementsProperties` structure from the provided YAML file.
+impl SupportedElements {
+    /// Construct a new `SupportedElements` structure from the provided YAML file.
     ///
     /// ## Returns
-    /// `ElementsProperties` structure if parsing was successful.
+    /// `SupportedElements` structure if parsing was successful.
     /// `ParseElementError` otherwise.
     ///
     /// ## Notes
     /// - For an example of the 'elements yaml file', see `src/config/elements.yaml`.
     pub fn new_from_file(filename: impl AsRef<Path>) -> Result<Self, ParseElementError> {
+        SupportedElements::new_from_string(&SupportedElements::load_yaml_to_string(filename)?)
+    }
+
+    /// Parse yaml string into `SupportedElements` structure.
+    fn new_from_string(yaml: &str) -> Result<Self, ParseElementError> {
+        let elements = parse_yaml_raw(yaml)?
+            .into_iter()
+            .map(|(name, temp)| temp.to_element(name))
+            .collect::<Result<Vec<Element>, ParseElementError>>()?;
+
+        Ok(SupportedElements { elements })
+    }
+
+    /// Update `SupportedElements` structure using data from the provided YAML file.
+    ///
+    /// ## Returns
+    /// `Ok` if the parsing was successful.
+    /// `ParseElementError` otherwise. If an error occurs, the `SupportedElements` structure is not changed.
+    ///
+    /// ## Example
+    /// Let's suppose that in the default `SupportedElements` structure provided by `groan_rs` library,
+    /// you are missing information about polonium
+    /// and also you don't like the default van der Waals radius for carbon.
+    /// You can construct a new yaml file containing the missing and modified information.
+    /// The yaml file, `my_elements.yaml`, should look like this:
+    /// ```yaml
+    /// ---
+    /// carbon:
+    ///   vdw: 0.20
+    /// polonium:
+    ///   symbol: Po
+    ///   query: name r'^[Pp][Oo].*$'
+    ///   mass: 209.0
+    /// ...
+    /// ```
+    ///
+    /// The modified `SupportedElements` structure can be obtained using:
+    /// ```no_run
+    /// use groan_rs::prelude::*;
+    ///
+    /// // load default parameters for the elements
+    /// let mut elements = SupportedElements::default();
+    /// // update the `elements` structure with your custom information
+    /// elements.update_from_file("my_elements.yaml").unwrap();
+    /// ```
+    ///
+    /// In the `elements` structure, carbon will now have `vdw` of 0.2,
+    /// and a new element, polonium, will be added.
+    /// No other information about the elements will be changed.
+    pub fn update_from_file(
+        &mut self,
+        filename: impl AsRef<Path>,
+    ) -> Result<(), ParseElementError> {
+        SupportedElements::update_from_string(
+            self,
+            &SupportedElements::load_yaml_to_string(filename)?,
+        )
+    }
+
+    /// Opens the specified file and loads its contents into a string.
+    fn load_yaml_to_string(filename: impl AsRef<Path>) -> Result<String, ParseElementError> {
         let mut yaml_file = match File::open(&filename) {
             Ok(x) => x,
             Err(_) => {
@@ -75,7 +188,7 @@ impl ElementsProperties {
 
         let mut yaml_string = String::new();
         match yaml_file.read_to_string(&mut yaml_string) {
-            Ok(_) => ElementsProperties::new_from_string(&yaml_string),
+            Ok(_) => Ok(yaml_string),
             Err(_) => {
                 return Err(ParseElementError::FileCouldNotBeRead(Box::from(
                     filename.as_ref(),
@@ -84,72 +197,152 @@ impl ElementsProperties {
         }
     }
 
-    /// Parse yaml string into `ElementsProperties` structure.
-    fn new_from_string(yaml: &str) -> Result<Self, ParseElementError> {
-        #[derive(Deserialize, Debug)]
-        #[serde(deny_unknown_fields)]
-        struct TempElement {
-            symbol: Option<String>,
-            query: Option<String>,
-            mass: Option<f32>,
-            vdw: Option<f32>,
-            expected_max_bonds: Option<u8>,
+    /// Update `SupportedElements` structure from yaml string.
+    /// Previously defined elements are updated. New elements are added.
+    /// In case an error occurs, `self` is not changed.
+    fn update_from_string(&mut self, yaml: &str) -> Result<(), ParseElementError> {
+        let parsed_elements = parse_yaml_raw(yaml)?;
+
+        // first check for the validity of all elements
+        // this is necessary because we do not want to disrupt the SupportedElements
+        // structure if the parsed yaml is invalid
+        for (name, temp) in &parsed_elements {
+            temp.validate(name)?;
         }
 
-        // if parsing yaml fails, return location of the error
-        let raw_elements: HashMap<String, TempElement> = match serde_yaml::from_str(yaml) {
-            Ok(x) => x,
-            Err(e) => {
-                let (line, col) = match e.location() {
-                    Some(x) => (x.line(), x.column()),
-                    None => (0, 0),
-                };
-
-                return Err(ParseElementError::CouldNotParseYaml(line, col));
+        // then update the structure
+        for (name, temp) in parsed_elements {
+            if let Some(element) = self.elements.iter_mut().find(|e| e.name == name) {
+                temp.update_element(element);
+            } else {
+                let new_element = temp.to_element(name)
+                    .expect("FATAL GROAN ERROR | SupportedElements::update_from_string | TempElement should be valid.");
+                self.elements.push(new_element);
             }
-        };
+        }
 
-        // convert temporary elements into final elements
-        // parse selection queries to select trees
-        // check for validity of the mass and vdw
-        let elements = raw_elements
-            .into_iter()
-            .map(|(name, element)| {
-                let select = match element.query {
-                    Some(x) => match parse_query(&x) {
-                        Ok(x) => Some(*x),
-                        Err(e) => return Err(ParseElementError::InvalidQuery(e)),
-                    },
-                    None => None,
-                };
-
-                match element.mass {
-                    Some(x) if x < 0.0 => return Err(ParseElementError::InvalidMass(name)),
-                    Some(_) | None => (),
-                }
-
-                match element.vdw {
-                    Some(x) if x < 0.0 => return Err(ParseElementError::InvalidVdW(name)),
-                    Some(_) | None => (),
-                }
-
-                Ok(Element {
-                    name,
-                    symbol: element.symbol,
-                    select,
-                    mass: element.mass,
-                    vdw: element.vdw,
-                    expected_max_bonds: element.expected_max_bonds,
-                })
-            })
-            .collect::<Result<Vec<Element>, ParseElementError>>()?;
-
-        Ok(ElementsProperties { elements })
+        Ok(())
     }
 }
 
+/// Temporary `Element` structure for parsing YAML.
+#[derive(Deserialize, Debug)]
+#[serde(deny_unknown_fields)]
+struct TempElement {
+    symbol: Option<String>,
+    query: Option<String>,
+    mass: Option<f32>,
+    vdw: Option<f32>,
+    expected_max_bonds: Option<u8>,
+}
+
+impl TempElement {
+    /// Convert `TempElement` and name to `Element` structure.
+    /// Checks for validity of the query, mass, and vdw radius.
+    #[allow(clippy::wrong_self_convention)]
+    fn to_element(self, name: String) -> Result<Element, ParseElementError> {
+        let select = TempElement::query2select(self.query.as_deref())?;
+
+        TempElement::check_mass(self.mass, &name)?;
+        TempElement::check_vdw(self.vdw, &name)?;
+
+        Ok(Element {
+            name,
+            symbol: self.symbol,
+            select,
+            mass: self.mass,
+            vdw: self.vdw,
+            expected_max_bonds: self.expected_max_bonds,
+        })
+    }
+
+    /// Update `Element` structure based on the provided `TempElement` structure.
+    ///
+    /// ## Panics
+    /// Panics if the `TempElement` is invalid (its validity should be checked before using `TempElement::validate`).
+    fn update_element(self, element: &mut Element) {
+        if self.symbol.is_some() {
+            element.symbol = self.symbol;
+        }
+
+        if self.query.is_some() {
+            element.select = TempElement::query2select(self.query.as_deref())
+                .expect("FATAL GROAN ERROR | TempElement::update_element | Query should be valid.");
+        }
+
+        if self.mass.is_some() {
+            TempElement::check_mass(self.mass, &element.name)
+                .expect("FATAL GROAN ERROR | TempElement::update_element | Mass should be valid.");
+            element.mass = self.mass;
+        }
+
+        if self.vdw.is_some() {
+            TempElement::check_vdw(self.vdw, &element.name)
+                .expect("FATAL GROAN ERROR | TempElement::update_element | VdW should be valid.");
+            element.vdw = self.vdw;
+        }
+
+        if self.expected_max_bonds.is_some() {
+            element.expected_max_bonds = self.expected_max_bonds;
+        }
+    }
+
+    /// Checks that the properties of the TempElement are valid.
+    fn validate(&self, name: &str) -> Result<(), ParseElementError> {
+        TempElement::query2select(self.query.as_deref())?;
+        TempElement::check_mass(self.mass, name)?;
+        TempElement::check_vdw(self.vdw, name)?;
+
+        Ok(())
+    }
+
+    /// Convert query to select tree. If `query` is None, returns `Ok(None)`.
+    fn query2select(query: Option<&str>) -> Result<Option<Select>, ParseElementError> {
+        match query {
+            Some(x) => match parse_query(x) {
+                Ok(x) => Ok(Some(*x)),
+                Err(e) => Err(ParseElementError::InvalidQuery(e)),
+            },
+            None => Ok(None),
+        }
+    }
+
+    /// Check that `mass` is not negative. If `mass` is None, returns `Ok`.
+    fn check_mass(mass: Option<f32>, name: &str) -> Result<(), ParseElementError> {
+        match mass {
+            Some(x) if x < 0.0 => Err(ParseElementError::InvalidMass(name.to_owned())),
+            Some(_) | None => Ok(()),
+        }
+    }
+
+    /// Check that `vdw` is not negative. If `mass` is None, returns `Ok`.
+    fn check_vdw(vdw: Option<f32>, name: &str) -> Result<(), ParseElementError> {
+        match vdw {
+            Some(x) if x < 0.0 => Err(ParseElementError::InvalidVdW(name.to_owned())),
+            Some(_) | None => Ok(()),
+        }
+    }
+}
+
+fn parse_yaml_raw(yaml: &str) -> Result<HashMap<String, TempElement>, ParseElementError> {
+    // if parsing yaml fails, return location of the error
+    let raw_elements: HashMap<String, TempElement> = match serde_yaml::from_str(yaml) {
+        Ok(x) => x,
+        Err(e) => {
+            let (line, col) = match e.location() {
+                Some(x) => (x.line(), x.column()),
+                None => (0, 0),
+            };
+
+            return Err(ParseElementError::CouldNotParseYaml(line, col));
+        }
+    };
+
+    Ok(raw_elements)
+}
+
 /******************************/
-/*   UNIT TESTS FOR ELEMENT   */
+/*         UNIT TESTS         */
 /******************************/
 
 #[cfg(test)]
@@ -161,7 +354,7 @@ mod tests {
 
     #[test]
     fn elements_default() {
-        let elements = ElementsProperties::default();
+        let elements = SupportedElements::default();
 
         assert_eq!(elements.elements.len(), 39);
 
@@ -173,6 +366,8 @@ mod tests {
         assert!(names.contains(&String::from("bismuth")));
         assert!(names.contains(&String::from("lithium")));
 
+        let mut visited = 0;
+
         for e in &elements.elements {
             // first element
             if e.name == "dummy" {
@@ -181,6 +376,8 @@ mod tests {
                 assert_approx_eq!(f32, e.mass.unwrap(), 0.0);
                 assert!(e.vdw.is_none());
                 assert!(e.expected_max_bonds.is_none());
+                visited += 1;
+
                 continue;
             }
 
@@ -191,6 +388,8 @@ mod tests {
                 assert_approx_eq!(f32, e.mass.unwrap(), 12.0107);
                 assert_approx_eq!(f32, e.vdw.unwrap(), 0.17);
                 assert_eq!(e.expected_max_bonds.unwrap(), 4);
+                visited += 1;
+
                 continue;
             }
 
@@ -201,33 +400,32 @@ mod tests {
                 assert_approx_eq!(f32, e.mass.unwrap(), 208.98040);
                 assert!(e.vdw.is_none());
                 assert!(e.expected_max_bonds.is_none());
+                visited += 1;
+
                 continue;
             }
         }
+
+        assert_eq!(visited, 3);
     }
 
-    #[test]
-    fn elements_from_file() {
-        let mut default_elements = ElementsProperties::default();
-        let mut elements_from_file =
-            ElementsProperties::new_from_file("src/config/elements.yaml").unwrap();
-
+    fn elements_match(el1: &mut SupportedElements, el2: &mut SupportedElements) {
         assert_eq!(
-            default_elements.elements.len(),
-            elements_from_file.elements.len()
+            el1.elements.len(),
+            el2.elements.len()
         );
 
-        default_elements
+        el1
             .elements
             .sort_by(|x, y| x.name.cmp(&y.name));
-        elements_from_file
+        el2
             .elements
             .sort_by(|x, y| x.name.cmp(&y.name));
 
-        for (e1, e2) in default_elements
+        for (e1, e2) in el1
             .elements
             .iter()
-            .zip(elements_from_file.elements.iter())
+            .zip(el2.elements.iter())
         {
             assert_eq!(e1.name, e2.name);
 
@@ -264,8 +462,17 @@ mod tests {
     }
 
     #[test]
+    fn elements_from_file() {
+        let mut default_elements = SupportedElements::default();
+        let mut elements_from_file =
+            SupportedElements::new_from_file("src/config/elements.yaml").unwrap();
+
+        elements_match(&mut default_elements, &mut elements_from_file);
+    }
+
+    #[test]
     fn elements_invalid_field() {
-        match ElementsProperties::new_from_file("test_files/elements_invalid_field.yaml") {
+        match SupportedElements::new_from_file("test_files/elements_invalid_field.yaml") {
             Ok(_) => panic!("Function should have failed but it succeeded."),
             Err(ParseElementError::CouldNotParseYaml(line, col)) => {
                 assert_eq!(line, 10);
@@ -280,7 +487,7 @@ mod tests {
 
     #[test]
     fn elements_nonexistent_file() {
-        match ElementsProperties::new_from_file("nonexistent_file.yaml") {
+        match SupportedElements::new_from_file("nonexistent_file.yaml") {
             Ok(_) => panic!("Function should have failed but it succeeded."),
             Err(ParseElementError::FileNotFound(name)) => {
                 assert_eq!(name.to_str().unwrap(), "nonexistent_file.yaml")
@@ -294,7 +501,7 @@ mod tests {
 
     #[test]
     fn elements_invalid_query() {
-        match ElementsProperties::new_from_file("test_files/elements_invalid_query.yaml") {
+        match SupportedElements::new_from_file("test_files/elements_invalid_query.yaml") {
             Ok(_) => panic!("Function should have failed but it succeeded."),
             Err(ParseElementError::InvalidQuery(SelectError::InvalidNumber(query))) => {
                 assert_eq!(query, "resid 14 15 17 X 18 19 20")
@@ -308,7 +515,7 @@ mod tests {
 
     #[test]
     fn elements_invalid_mass() {
-        match ElementsProperties::new_from_file("test_files/elements_invalid_mass.yaml") {
+        match SupportedElements::new_from_file("test_files/elements_invalid_mass.yaml") {
             Ok(_) => panic!("Function should have failed but it succeeded."),
             Err(ParseElementError::InvalidMass(name)) => assert_eq!(name, String::from("hydrogen")),
             Err(e) => panic!(
@@ -320,7 +527,7 @@ mod tests {
 
     #[test]
     fn elements_invalid_vdw() {
-        match ElementsProperties::new_from_file("test_files/elements_invalid_vdw.yaml") {
+        match SupportedElements::new_from_file("test_files/elements_invalid_vdw.yaml") {
             Ok(_) => panic!("Function should have failed but it succeeded."),
             Err(ParseElementError::InvalidVdW(name)) => assert_eq!(name, String::from("carbon")),
             Err(e) => panic!(
@@ -329,4 +536,171 @@ mod tests {
             ),
         }
     }
+
+    #[test]
+    fn elements_update() {
+        let mut elements = SupportedElements::default();
+        elements
+            .update_from_file("test_files/elements_update.yaml")
+            .unwrap();
+
+        let mut visited = 0;
+
+        for e in &elements.elements {
+            // first element
+            if e.name == "dummy" {
+                assert!(e.symbol.is_none());
+                assert!(e.select.is_some());
+                assert_approx_eq!(f32, e.mass.unwrap(), 0.0);
+                assert!(e.vdw.is_none());
+                assert!(e.expected_max_bonds.is_none());
+
+                visited += 1;
+                continue;
+            }
+
+            // changed element #1
+            if e.name == "carbon" {
+                assert_eq!(e.symbol.as_ref().unwrap(), "C");
+                assert!(e.select.is_some());
+                assert_approx_eq!(f32, e.mass.unwrap(), 12.0107);
+                assert_approx_eq!(f32, e.vdw.unwrap(), 0.20);
+                assert_eq!(e.expected_max_bonds.unwrap(), 4);
+
+                visited += 1;
+                continue;
+            }
+
+            // change element #2
+            if e.name == "hydrogen" {
+                assert_eq!(e.symbol.as_ref().unwrap(), "HH");
+                assert!(e.select.is_some());
+                assert_approx_eq!(f32, e.mass.unwrap(), 2.014);
+                assert_approx_eq!(f32, e.vdw.unwrap(), 0.12);
+                assert_eq!(e.expected_max_bonds.unwrap(), 8);
+
+                visited += 1;
+                continue;
+            }
+
+            // added element
+            if e.name == "polonium" {
+                assert_eq!(e.symbol.as_ref().unwrap(), "Po");
+                assert!(e.select.is_some());
+                assert_approx_eq!(f32, e.mass.unwrap(), 209.0);
+                assert!(e.vdw.is_none());
+                assert!(e.expected_max_bonds.is_none());
+
+                visited += 1;
+                continue;
+            }
+
+            // last element
+            if e.name == "bismuth" {
+                assert_eq!(e.symbol.as_ref().unwrap(), "Bi");
+                assert!(e.select.is_some());
+                assert_approx_eq!(f32, e.mass.unwrap(), 208.98040);
+                assert!(e.vdw.is_none());
+                assert!(e.expected_max_bonds.is_none());
+
+                visited += 1;
+                continue;
+            }
+        }
+
+        assert_eq!(visited, 5);
+    }
+
+    #[test]
+    fn elements_update_invalid_field() {
+        let mut elements = SupportedElements::default();
+
+        match elements.update_from_file("test_files/elements_invalid_field.yaml") {
+            Ok(_) => panic!("Function should have failed but it succeeded."),
+            Err(ParseElementError::CouldNotParseYaml(line, col)) => {
+                assert_eq!(line, 10);
+                assert_eq!(col, 3);
+            }
+            Err(e) => panic!(
+                "Function failed successfully but incorrect error type `{}` was returned.",
+                e
+            ),
+        }
+
+        let mut default = SupportedElements::default();
+        elements_match(&mut elements, &mut default);
+    }
+
+    #[test]
+    fn elements_update_nonexistent_file() {
+        let mut elements = SupportedElements::default();
+
+        match elements.update_from_file("nonexistent_file.yaml") {
+            Ok(_) => panic!("Function should have failed but it succeeded."),
+            Err(ParseElementError::FileNotFound(name)) => {
+                assert_eq!(name.to_str().unwrap(), "nonexistent_file.yaml")
+            }
+            Err(e) => panic!(
+                "Function failed successfully but incorrect error type `{}` was returned.",
+                e
+            ),
+        }
+
+        let mut default = SupportedElements::default();
+        elements_match(&mut elements, &mut default);
+    }
+
+    #[test]
+    fn elements_update_invalid_query() {
+        let mut elements = SupportedElements::default();
+
+        match elements.update_from_file("test_files/elements_invalid_query.yaml") {
+            Ok(_) => panic!("Function should have failed but it succeeded."),
+            Err(ParseElementError::InvalidQuery(SelectError::InvalidNumber(query))) => {
+                assert_eq!(query, "resid 14 15 17 X 18 19 20")
+            }
+            Err(e) => panic!(
+                "Function failed successfully but incorrect error type `{}` was returned.",
+                e
+            ),
+        }
+
+        let mut default = SupportedElements::default();
+        elements_match(&mut elements, &mut default);
+    }
+
+    #[test]
+    fn elements_update_invalid_mass() {
+        let mut elements = SupportedElements::default();
+
+        match SupportedElements::new_from_file("test_files/elements_invalid_mass.yaml") {
+            Ok(_) => panic!("Function should have failed but it succeeded."),
+            Err(ParseElementError::InvalidMass(name)) => assert_eq!(name, String::from("hydrogen")),
+            Err(e) => panic!(
+                "Function failed successfully but incorrect error type `{}` was returned.",
+                e
+            ),
+        }
+
+        let mut default = SupportedElements::default();
+        elements_match(&mut elements, &mut default);
+    }
+
+    #[test]
+    fn elements_update_invalid_vdw() {
+        let mut elements = SupportedElements::default();
+
+        match SupportedElements::new_from_file("test_files/elements_invalid_vdw.yaml") {
+            Ok(_) => panic!("Function should have failed but it succeeded."),
+            Err(ParseElementError::InvalidVdW(name)) => assert_eq!(name, String::from("carbon")),
+            Err(e) => panic!(
+                "Function failed successfully but incorrect error type `{}` was returned.",
+                e
+            ),
+        }
+
+        let mut default = SupportedElements::default();
+        elements_match(&mut elements, &mut default);
+    }
+
 }