+++ changelog.md
@@ -1,6 +1,25 @@
 
 ## Changelog for the `groan_rs` library
 
+### Version 0.6.0
+
+#### `Element` and `SupportedElements` structures
+- `Element` structure defines properties of an element, such as its name, symbol, mass, van der Waals radius etc.
+- `SupportedElements` structure contains all supported elements.
+- `SupportedElements` structure can be constructed from a YAML file containing the element definitions by calling `SupportedElements::new_from_file`.
+- Default properties of the elements supported by `groan_rs` are stored in `src/config/elements.yaml`. Default `SupportedElements` structure, containing elements and their properties from this YAML file, can be constructed using `SupportedElements::default`.
+- `SupportedElements` structure can be updated using `SupportedElements::update_from_file`. In such case, a new YAML file must be provided which specifies changes to be made in the `SupportedElements` structure.
+
+#### Changes to the `Atom` structure
+- New optional fields `mass` and `element` added:
+  - `mass` specifies mass of the atom.
+  - `element` specifies the element of the atom. The element also contains information about the mass. 
+  If the atom has no explicitely provided mass but does have an element assigned, the mass of the element
+  is used as the mass of the atom.
+- Methods were added to use and access `mass` and `element`.
+
+***
+
 ### Version 0.5.0
 #### Important BREAKING CHANGES affecting the entire `groan_rs` library